TASK: `resource:clean` abort if `Neos.Media` is not installed

+++ Neos.Flow/Classes/Command/ResourceCommandController.php
@@ -206,8 +206,6 @@ class ResourceCommandController extends CommandController
         $this->outputLine('Checking if resource data exists for all known resource objects ...');
         $this->outputLine();
 
-        $mediaPackagePresent = $this->packageManager->isPackageAvailable('Neos.Media');
-
         $resourcesCount = $this->resourceRepository->countAll();
         $this->output->progressStart($resourcesCount);
 
@@ -229,10 +227,12 @@ class ResourceCommandController extends CommandController
         $this->output->progressFinish();
         $this->outputLine();
 
-        /* @var AssetRepository $assetRepository */
-        $assetRepository = $this->objectManager->get(AssetRepository::class);
-        /* @var ThumbnailRepository $thumbnailRepository */
-        $thumbnailRepository = $this->objectManager->get(ThumbnailRepository::class);
+        // FIXME flow has no dependency on Neos.Media. This code should be extracted.
+        /* @var AssetRepository|null $assetRepository */
+        $assetRepository = class_exists(AssetRepository::class) ? $this->objectManager->get(AssetRepository::class) : null;
+        /* @var ThumbnailRepository|null $thumbnailRepository */
+        $thumbnailRepository = class_exists(ThumbnailRepository::class) ? $this->objectManager->get(ThumbnailRepository::class) : null;
+        $mediaPackagePresent = $assetRepository && $thumbnailRepository && $this->packageManager->isPackageAvailable('Neos.Media');
 
         if (count($brokenResources) > 0) {
             foreach ($brokenResources as $key => $resourceIdentifier) {
@@ -288,16 +288,18 @@ class ResourceCommandController extends CommandController
                         $resource->disableLifecycleEvents();
                         /** @psalm-suppress PossiblyNullArgument */
                         $this->resourceRepository->remove($resource);
-                        if (isset($relatedAssets[$resource])) {
-                            foreach ($relatedAssets[$resource] as $asset) {
-                                $assetRepository->removeWithoutUsageChecks($asset);
-                                $brokenAssetCounter++;
+                        if ($mediaPackagePresent) {
+                            if (isset($relatedAssets[$resource])) {
+                                foreach ($relatedAssets[$resource] as $asset) {
+                                    $assetRepository->removeWithoutUsageChecks($asset);
+                                    $brokenAssetCounter++;
+                                }
                             }
-                        }
-                        if (isset($relatedThumbnails[$resource])) {
-                            foreach ($relatedThumbnails[$resource] as $thumbnail) {
-                                $thumbnailRepository->remove($thumbnail);
-                                $brokenThumbnailCounter++;
+                            if (isset($relatedThumbnails[$resource])) {
+                                foreach ($relatedThumbnails[$resource] as $thumbnail) {
+                                    $thumbnailRepository->remove($thumbnail);
+                                    $brokenThumbnailCounter++;
+                                }
                             }
                         }
                         $this->persistenceManager->persistAll();
